Task REST added

+++ src/main/java/pl/power/services/impl/DefaultTaskService.java
@@ -4,14 +4,21 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
+import pl.power.domain.entities.PowerStation;
+import pl.power.domain.entities.Task;
 import pl.power.domain.entities.enums.TaskType;
+import pl.power.domain.repositories.PowerStationRepository;
 import pl.power.domain.repositories.TaskRepository;
+import pl.power.dtos.CreateTaskDTO;
 import pl.power.dtos.TaskDTO;
 import pl.power.services.TaskService;
 import pl.power.services.errors.IdIsNullException;
+import pl.power.services.errors.PowerStationsNotFoundException;
+import pl.power.services.errors.TaskNotFoundException;
 
 import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -20,22 +27,32 @@ import java.util.stream.Collectors;
 public class DefaultTaskService implements TaskService {
 
     private final TaskRepository taskRepository;
+    private final PowerStationRepository powerStationRepository;
     private final ModelMapper mapper;
 
     @Override
     @Transactional
-    public TaskDTO addTask(TaskDTO taskDTO) {
-        return null;
+    public Long add(CreateTaskDTO createTaskDTO) {
+        Task task = mapper.map(createTaskDTO, Task.class);
+        Optional<PowerStation> powerStationOptional = powerStationRepository.findById(task.getId());
+        PowerStation powerStation = powerStationOptional.orElseThrow(() -> new PowerStationsNotFoundException(task.getId()));
+        task.setId(null);
+        powerStation.add(task);
+        powerStationRepository.save(powerStation);
+        return taskRepository.findLastSaved();
     }
 
     @Override
     @Transactional
-    public void deleteTask(Long id) {
+    public void delete(Long id) {
+        if (id == null) {
+            throw new IdIsNullException();
+        }
         taskRepository.deleteById(id);
     }
 
     @Override
-    public List<TaskDTO> findAllTasks() {
+    public List<TaskDTO> findAll() {
         return taskRepository.findAllOneSelect()
                 .stream()
                 .map(task -> {
@@ -48,7 +65,12 @@ public class DefaultTaskService implements TaskService {
 
     @Override
     public TaskDTO findById(Long id) {
-        return null;
+        if (id == null) {
+            throw new IdIsNullException();
+        }
+        Task task = taskRepository.findById(id)
+                .orElseThrow(() -> new TaskNotFoundException(id));
+        return mapper.map(task, TaskDTO.class);
     }
 
     @Override
@@ -62,5 +84,19 @@ public class DefaultTaskService implements TaskService {
                 .filter(task -> task.getPowerStation().getId().equals(id))
                 .filter(value -> value.getTaskType() == filter)
                 .count();
+    }
+
+    @Override
+    @Transactional
+    public TaskDTO update(CreateTaskDTO createTaskDTO) {
+        Optional<Task> byId = taskRepository.findById(createTaskDTO.getId());
+        Task task = byId.orElseThrow(() -> new TaskNotFoundException(createTaskDTO.getId()));
+        task.setPowerLoss(createTaskDTO.getPowerLoss());
+        task.setTaskType(createTaskDTO.getTaskType());
+        task.setStartDate(createTaskDTO.getStartDate());
+        task.setEndDate(createTaskDTO.getEndDate());
+        Task save = taskRepository.save(task);
+        mapper.map(save,TaskDTO.class);
+        return mapper.map(save,TaskDTO.class);
     }
 }
